requirements.txt
streamlit==1.22.0
tensorflow==2.15.0
numpy==1.24.3
opencv-python==4.7.0.72
matplotlib==3.7.1
pillow==9.5.0
seaborn==0.12.2
scikit-learn==1.2.2
gdown==4.7.1
fpdf2==3.0.0
reportlab==4.0.4
pandas==2.0.3
plotly==5.15.0
